requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-JWT-Extended
passlib
python-dotenv
pymysql
werkzeug
openai
